New Branch: kwallet-dev.  Applied alpine-2.00-kwallet patch. This commit adds the kwallet-dev branch, and commits the initial application of the patch I created against the alpine-2.00 release.  I haven't run a test compile, but will shortly.
The major overview: kwallet.cpp is the main event.  It interfaces
with the C++ KDE libraries (hence the separate file).  imap.c
and imap.h were modified to take advantage of kwallet.cpp.  These
three files constitute the core of the kwallet password cache
feature.

alpine/Makefile.am and configure.ac have been updated to add
kwallet.cpp to the buildsystem with the appropriate compiler flags.

The rest of the changes are to headers.  alpine uses a number of
C++ reserved keywords in its datastructures, which while not a
problem for C, means the headers cannot be included as-is in C++
code.  Since kwallet.cpp is the only C++ code in the project, and
it doesn't use any of these datastructures, I added some #defines
conditional on the __cplusplus macro to work around this.  The C
files see unmodified header files and the C++ code sees headers
with alternate, unconflicting names.