changes configure.ac to detect Qt and KDE libs in a more portable manner The main action in this commit is in configure.ac.  I updated the compile variable detection for the kwallet feature to be more portable.  This was more complex than I expected--I'm glad the Subversion folks had already figured it out.
I also changed how the options for password caching work.  As suggested
by one of the comments in configure.ac, I broke out each password cache
into its own --with-* switch.  This primarily allows the user to specify
the KDE root with the --with-kwallet switch (which would not be possible
using the previous --with-local-password-cache-method switch).

Finally, I stubbed out Gnome Keyring options in configure.ac.  The macro
to add --with-gnome-keyring to the ./configure script is commented out,
but much of the rest isn't (and has associated error messages).

Tested by running through full compile, which completed successfully.